Handle errors, improve logging and cleanup on shutdown.
Rework error handling so that all write errors are processed instead of being ignored. This changes the flow of most handle functions to fail fast, with the natural flow being the none error case making it much easier to follow. This also removes the need for panic in handlePASV.

go-kit's logger requires a high number of additional vendor packages just to provide structured logging, so swap this out for logrus which is the most widely used logging package.

Ensure that client connections and signal handler are cleaned up on server shutdown.

Also:
* Improved segregation between server and client.
* Fixed nill driver check in handlePASS.
* Stripped \r\n from command lines to improve logging.
* Enabled keepalives on all connections to prevent deadlock on reads.
* Enabled dial timeouts for active connections.
* Various style cleanups.
* Improved localization of variable calculations.
* Corrected some incorrect comments.
* Removed needless conversions in SetDeadlne call.
* Converted to static commandsMap instead of init().
* Fixed variable shadowing.
* Remove unnecessary testing package.
* Fixed all issues identified by gometalinter.
* Eliminated unneeded buffered writer, which was only doing line buffering, which is very similar to the default anyway; eliminating the need for constant Flush calls.
* Enabled gometalinter.v2 checks under travis, removing redundant golint checks.