transfer open: add error checking
send StatusFileStatusOK only if transfer.Open() returns no error.
If an error happen send a 425 error code.

This patch fixes the following use case:

1) start the ftpserver as non root user
2) start an active connection
3) request file list

if the server is not started as root and ActiveTransferPortNon20 is not
set then the server cannot bind to port 20.

Before the patch this response is sent anyway:

c.writeMessage(StatusFileStatusOK, "Using transfer connection")

and the client hangs forever.

After this patch the ftp client receive this error:

ftp> ls
200 PORT command successful
425 Unable to open transfer: could not establish active connection due: dial tcp :20->127.0.0.1:37811: bind: permission denied

tested using ftp CLI on Linux