data channel: ensure to perform at least a write
if there is nothing to write to the data channel, for example an empty file,
or an empty file list, in TLS mode we close the connection before that the
TLS handshake happens and so the clients will complaints.

This patch write an empty buffer, if there is nothing to send, to avoid this
issue.

Fixes #159