ensure to always close the control connection
if ClientConnected returns an error, the control connection is not
closed so a client could keep it open indefinitely.

Closing the control connection within the end() method ensures that
the connection is always closed.

If the connection was already closed client side we get the error
"use of closed network connection", this error should not cause
any issue