add an optional interface to wrap the passive connections listener
If we allow a trusted network when using HAProxy we don't have any
security. All the passive connections will come through the proxy
and so they will be allowed.

We have to require to add the proxy header for both control and passive
data connection. The new optional interface allows to wrap the passive
listener and so decode the proxy header and get the real client IP